initial

+++ Smaller Sum.cpp
@@ -11,16 +11,16 @@ public:
     vector<long long> smallerSum(int n, vector<int> &arr)
     {
         // Code here
-        vector<vector<ll>> v;
+        vector<vector<long long>> v;
         for (int i = 0; i < n; i++)
-            v.push_back({i, (ll)arr[i]});
+            v.push_back({i, (long long)arr[i]});
 
-        sort(v.begin(), v.end(), [](const vector<ll> &a, const vector<ll> &b)
+        sort(v.begin(), v.end(), [](const vector<long long> &a, const vector<long long> &b)
              { return a[1] < b[1]; });
 
         for (int i = 1; i < n; i++)
             v[i][1] += v[i - 1][1];
-        vector<ll> res(n, 0);
+        vector<long long> res(n, 0);
 
         res[v[0][0]] = 0;
         int idx = -1;
